requirements.txt
Django==2.1.3
djangorestframework==3.9.0
drf-renderer-xlsx==0.3.0
et-xmlfile==1.0.1
jdcal==1.4
mysqlclient==1.3.13
openpyxl==2.5.10
